+++ react/components/Box/index.js
@@ -2,7 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 const Box = ({ children }) => (
-  <div className="bg-base pa7 br3 b--muted-4 ba">{children}</div>
+  <div className="bg-base t-body c-on-base pa7 br3 b--muted-4 ba">
+    {children}
+  </div>
 )
 
 Box.propTypes = {