+++ react/components/PageHeader/index.js
@@ -34,7 +34,7 @@ class PageHeader extends PureComponent {
         )}
 
         <div
-          className={`vtex-pageHeader__title c-on-base f2 fw6 flex flex-row
+          className={`vtex-pageHeader__title c-on-base t-heading-2 flex flex-row
           flex-wrap justify-between ${linkLabel ? 'mt0' : 'mt7'}`}>
           <span>{this.props.title}</span>