+++ react/components/Pagination/index.js
@@ -52,9 +52,11 @@ class Pagination extends PureComponent {
         }`}>
         {dropdownOptions && (
           <div className="flex flex-row pt5 items-baseline">
-            <span className="mr4 c-muted-2 f6">{this.props.textShowRows}</span>
+            <span className="mr4 c-muted-2 t-small">
+              {this.props.textShowRows}
+            </span>
             <Dropdown
-              label=""
+              size="small"
               options={dropdownOptions}
               value={dropdownOptions[selectedRowsOptionIndex].label}
               onChange={this.handleRowsChange}
@@ -63,7 +65,7 @@ class Pagination extends PureComponent {
         )}
 
         <div className="flex flex-row pt5 items-center">
-          <div className="c-muted-2 f6">
+          <div className="c-muted-2 t-small">
             {this.props.currentItemFrom}
             {' - '}
             {this.props.currentItemTo} {this.props.textOf}{' '}