+++ react/components/Table/Toolbar.js
@@ -143,7 +143,7 @@ class Toolbar extends PureComponent {
                 variation="tertiary"
                 size="small"
                 onClick={() => this.handleToggleBox('isDensityBoxVisible')}>
-                <span className="flex align-baseline items-center near-black">
+                <span className="flex align-baseline items-center c-on-base">
                   <span className="mr3">
                     <IconDensity color="currentColor" />
                   </span>
@@ -154,7 +154,7 @@ class Toolbar extends PureComponent {
                 <div
                   className={`absolute ${
                     density.alignMenu === 'right' ? 'right-0' : 'left-0'
-                  } z-999 ba b--light-gray br2 mt2`}
+                  } z-999 ba b--muted-4 br2 mt2`}
                   style={BOX_SHADOW_STYLE}>
                   <div className="w-100 b2 br2 bg-base">
                     <div
@@ -167,7 +167,7 @@ class Toolbar extends PureComponent {
                             key={index}
                             className={`flex justify-between ph6 pv3 ${
                               isKeySelected ? 'b--emphasis' : 'b--transparent'
-                            } pointer hover-bg-light-silver bl bw1`}
+                            } pointer hover-bg-muted-5 bl bw1`}
                             onClick={() => {
                               handleToggleDensity(key)
                               this.handleToggleBox('isDensityBoxVisible')
@@ -194,7 +194,7 @@ class Toolbar extends PureComponent {
                 variation="tertiary"
                 size="small"
                 onClick={() => this.handleToggleBox('isFieldsBoxVisible')}>
-                <span className="flex align-baseline near-black">
+                <span className="flex align-baseline c-on-base">
                   <span className="mr3">
                     <IconColumns color="currentColor" />
                   </span>
@@ -205,14 +205,14 @@ class Toolbar extends PureComponent {
                 <div
                   className={`absolute ${
                     fields.alignMenu === 'right' ? 'right-0' : 'left-0'
-                  } z-999 ba b--light-gray br2 mt2`}>
+                  } z-999 ba b--muted-4 br2 mt2`}>
                   <div
                     className="w-100 b2 br2 bg-base"
                     style={{
                       ...BOX_SHADOW_STYLE,
                       width: FIELDS_BOX_WIDTH,
                     }}>
-                    <div className="flex inline-flex bb b--light-gray w-100 pl6 pv4">
+                    <div className="flex inline-flex bb b--muted-4 w-100 pl6 pv4">
                       <Button
                         variation="secondary"
                         size="small"
@@ -234,7 +234,7 @@ class Toolbar extends PureComponent {
                       {Object.keys(schema.properties).map((field, index) => (
                         <div
                           key={index}
-                          className="flex justify-between ph6 pv3 pointer hover-bg-light-silver"
+                          className="flex justify-between ph6 pv3 pointer hover-bg-muted-5"
                           onClick={() => toggleColumn(field)}>
                           <span className="w-70 truncate">
                             {schema.properties[field].title || field}
@@ -256,7 +256,7 @@ class Toolbar extends PureComponent {
               variation="tertiary"
               size="small"
               onClick={download.handleCallback}>
-              <span className="flex align-baseline near-black">
+              <span className="flex align-baseline c-on-base">
                 <span className="mr3">
                   <IconDownload color="currentColor" />
                 </span>
@@ -269,7 +269,7 @@ class Toolbar extends PureComponent {
               variation="tertiary"
               size="small"
               onClick={upload.handleCallback}>
-              <span className="flex align-baseline near-black">
+              <span className="flex align-baseline c-on-base">
                 <span className="mr3">
                   <IconUpload color="currentColor" />
                 </span>
@@ -288,7 +288,7 @@ class Toolbar extends PureComponent {
                 onClick={() =>
                   this.handleToggleBox('isExtraActionsBoxVisible')
                 }>
-                <span className="flex align-baseline items-center near-black">
+                <span className="flex align-baseline items-center c-on-base">
                   <span className="mr3">{extraActions.label}</span>
                   <IconCaretDown height={13} color="currentColor" />
                 </span>
@@ -297,7 +297,7 @@ class Toolbar extends PureComponent {
                 <div
                   className={`absolute ${
                     extraActions.alignMenu === 'left' ? 'left-0' : 'right-0'
-                  } z-999 ba b--light-gray br2`}
+                  } z-999 ba b--muted-4 br2`}
                   style={BOX_SHADOW_STYLE}>
                   <div
                     className="w-100 b2 br2 bg-base"
@@ -308,7 +308,7 @@ class Toolbar extends PureComponent {
                       {extraActions.actions.map((action, index) => (
                         <div
                           key={index}
-                          className="flex justify-between ph6 pv3 pointer hover-bg-light-silver"
+                          className="flex justify-between ph6 pv3 pointer hover-bg-muted-5"
                           onClick={action.handleCallback}>
                           <span className="w-70 truncate">{action.label}</span>
                         </div>